refactor: Move subscription to response handler
