+++ src/app/interactions/componentInteractions/handlers/configuration/2.2_handleRequestApiKeys.ts
@@ -1,13 +1,9 @@
-import { wsChain } from '@api/gqlClients';
 import { insertCircleApiTokens } from '@api/insertCircleApiTokens';
 import { ButtonStyle, ComponentButton, ComponentContext, ComponentType } from 'slash-create';
 import { CustomId } from 'src/app/interactions/customId';
 import { DiscordService } from 'src/app/service/DiscordService';
 import { extractCircleId } from 'src/app/utils/extractCircleId';
-import { isMessage } from 'src/app/utils/isMessage';
-import { sleep } from 'src/app/utils/sleep';
 import { disableAllComponents } from '../common';
-import { handleSendAlerts } from './3_handleSendAlerts';
 
 /**
  * Request API keys flow
@@ -40,34 +36,10 @@ export async function handleRequestApiKeys(ctx: ComponentContext) {
 		custom_id: CustomId.AuthorizeLinkCircleButton,
 	};
 
-	const message = await ctx.send({
+	await ctx.send({
 		content: `In order for discord to detect your authorization please click the button below, then go to the following link https://coordinape-git-staging-coordinape.vercel.app/discord/link?id=${rowId}&circleId=${circleId}`,
 		components: [
 			{ type: ComponentType.ACTION_ROW, components: [AUTHORIZE_LINK_CIRCLE_BUTTON] },
 		],
-	});
-
-	const onDiscordCircleApiToken = wsChain('subscription')({
-		discord_circle_api_tokens: [
-			{ where: { circle_id: { _eq: circleId } } },
-			{ circle_id: true, token: true },
-		],
-	});
-	
-	onDiscordCircleApiToken.on(async ({ discord_circle_api_tokens }) => {
-		const circleApiToken = discord_circle_api_tokens.find(({ circle_id }) => circle_id === circleId);
-		if (circleApiToken && circleApiToken.token) {
-			onDiscordCircleApiToken.ws.close();
-			if (isMessage(message)) {
-				message.edit({ components: [{ type: ComponentType.ACTION_ROW, components: [{ ...AUTHORIZE_LINK_CIRCLE_BUTTON, disabled: true }] }] });
-			}
-			await ctx.send({ content: `<@${ctx.user.id}>, you've linked the circle successfully!` });
-
-			// Just to improve message flow
-			await sleep(3000);
-
-			// Next question
-			return handleSendAlerts(ctx);
-		}
 	});
 }
